refactor: 메인페이지 캐러셀, 스테이지 컴포넌트 리팩토링
