+++ src/features/event/components/stage/MyStage.tsx
@@ -3,8 +3,7 @@
 import { useRouter } from 'next/navigation'
 import { useEffect, useState } from 'react'
 
-import { DUMMY_POSTER_DATA } from '@/constants/dummy.ts'
-import Card from '@/features/event/components/stage/Card/index.tsx'
+import StageList from '@/features/event/components/stage/StageList.tsx'
 import { StageCategory } from '@/features/event/types/StageCategory.ts'
 
 const USER = {
@@ -15,10 +14,6 @@ export default function MyStage() {
   const router = useRouter()
   const [isAuthenticated, setIsAuthenticated] = useState(false)
   const [category, setCategory] = useState<StageCategory>('all')
-
-  const handleClick = (cat: StageCategory) => {
-    setCategory(cat)
-  }
 
   useEffect(() => {
     const authStatus = localStorage.getItem('isAuthenticated') === 'true'
@@ -45,66 +40,14 @@ export default function MyStage() {
           <button
             type="button"
             className="text-disabled text-base font-medium px-4 py-1 rounded-xl leading-6"
+            onClick={() => router.push('/')}
           >
             더보기 &gt;
           </button>
         </h2>
       </div>
       {isAuthenticated ? (
-        <>
-          <div className="w-full flex space-x-4 justify-start text-xl mt-5">
-            <button
-              type="button"
-              onClick={() => handleClick('all')}
-              className={`${category === 'all' ? 'bg-primary border border-primary' : 'border border-border'} py-2 px-6 rounded-full min-w-max`}
-            >
-              전체
-            </button>
-            <button
-              type="button"
-              onClick={() => handleClick('performance')}
-              className={`${category === 'performance' ? 'bg-primary border border-primary' : 'border border-border'} py-2 px-6 rounded-full min-w-max`}
-            >
-              공연
-            </button>
-            <button
-              type="button"
-              onClick={() => handleClick('lecture')}
-              className={`${category === 'lecture' ? 'bg-primary border border-primary' : 'border border-border'} py-2 px-6 rounded-full min-w-max`}
-            >
-              강연
-            </button>
-            <button
-              type="button"
-              onClick={() => handleClick('club')}
-              className={`${category === 'club' ? 'bg-primary border border-primary' : 'border border-border'} py-2 px-6 rounded-full min-w-max`}
-            >
-              소모임
-            </button>
-            <button
-              type="button"
-              onClick={() => handleClick('etc')}
-              className={`${category === 'etc' ? 'bg-primary border border-primary' : 'border border-border'} py-2 px-6 rounded-full min-w-max`}
-            >
-              기타
-            </button>
-          </div>
-          <div className="mt-9 flex flex-wrap items-center justify-between gap-y-8 gap-x-3">
-            {DUMMY_POSTER_DATA.map((data) => {
-              return (
-                <Card
-                  key={data.title}
-                  title={data.title}
-                  isEnd={data.isEnd}
-                  location={data.location}
-                  sDate={data.startDate}
-                  content={data.content}
-                  posterUrl={data.posterUrl}
-                />
-              )
-            })}
-          </div>
-        </>
+        <StageList category={category} setCategory={setCategory} />
       ) : (
         <div className="w-full flex bg-yellow bg-opacity-40 justify-center items-center rounded-xl">
           <div className="flex flex-col items-center py-10 gap-5">
